code/requirements.txt
# django packages
Django==5.1.7
django-environ==0.12.0
django-cors-headers==4.7.0
djangorestframework==3.15.2

# api
drf-spectacular==0.28.0
coreapi==2.3.3

# large-scale data processing
polars[pandas]==1.27.1

# deduce packages
deduce==3.0.3
